+++ CHANGELOG.md
@@ -7,3 +7,4 @@
 - [Optimize: Maybe remove Chinese characters](https://github.com/Tencent/spring-cloud-tencent/pull/609)
 - [Bugfix: fix feign report call result error when using feign direct call](https://github.com/Tencent/spring-cloud-tencent/pull/623)
 - [optimize:optimize PolarisRouterContext and constants.](https://github.com/Tencent/spring-cloud-tencent/pull/628)
+- [support spring-retry router](https://github.com/Tencent/spring-cloud-tencent/pull/631)